backend/requirements.txt
Django==4.2.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
python-dateutil==2.8.2
Pillow==10.1.0
openpyxl==3.1.2
